[cs] refactor 'lp' to 'sl1m'
